Reorganize Gaia + GaiaSession for  more clarity
This commit simplifies and separates the duties of Gaia and GaiaSession, and improves readability. While the previous implementation was more 1:1 with the js api, there are redudant classes (i.e. Gaia, which does nothing but store a reference to a GaiaSession), uncertain states (you can have a GaiaSession, but it may or may not have a valid config), and unexpected side-effects (connecting to a GaiaSession) also stores its settings on disk as the default.

The new structure is as so:
- GaiaSession has been renamed to GaiaHubSession, as it represents a connection to a specific GaiaHub.
- It is not possible to instantiate a GaiaHubSession without having a valid config first, ensuring the object is valid.
- The only functionality a GaiaHubSession allows is getting and putting to that GaiaHub.
- Gaia is a class with all static members that manages connecting to hubs, storing default settings for later use, or managing multiple hubs (once we have that functionality on the platform).